feat(runtime): sync whole-GB RAM instead of live load sampling
Live CPU/RAM/GPU utilization is no longer collected or synced: the
burst sampler command and its perf-utils module are removed (the
machine_label helper moves to its own module), and the status push
carries only machine identity + lifetime session census. Total RAM
now ships as an approximate whole number (31.6 -> 32 GB), with a
cheap sysinfo fallback when hardware detection degrades. Roster
cards show a 'RAM 32 GB' chip plus GPU name; cpu/gpu locale keys
dropped across all 13 languages. Server counterpart: migration 0011
(upsert ignores legacy sample keys, so old clients stay compatible).
