+++ src/api/tauri/perf/metrics.ts
@@ -10,7 +10,6 @@ import type {
   ProcessMetrics,
   SystemInfo,
   SystemMemoryMetrics,
-  SystemRuntimeSnapshot,
 } from "./types";
 
 export async function getProcessMetrics(): Promise<ProcessMetrics> {
@@ -31,9 +30,4 @@ export async function detectLocalModelHardware(): Promise<LocalModelHardwareSumm
 
 export async function getSystemInfo(): Promise<SystemInfo> {
   return invoke<SystemInfo>("get_system_info");
-}
-
-/** ~1–2s whole-machine CPU/mem/GPU burst sample (member-runtime sharing). */
-export async function systemRuntimeSnapshot(): Promise<SystemRuntimeSnapshot> {
-  return invoke<SystemRuntimeSnapshot>("system_runtime_snapshot");
 }
