+++ src/features/Org2Cloud/memberRuntime/memberRuntimePushScheduler.test.ts
@@ -154,14 +154,6 @@ function makeDeps(
       chipType: "Apple M3",
       appVersion: "1.0.0",
     }),
-    getSample: vi.fn().mockResolvedValue({
-      cpuPercent: 10,
-      memUsedMb: 1000,
-      memTotalMb: 2000,
-      gpuPercent: null,
-      sampledOverMs: 1000,
-      sampledAtMs: NOW,
-    }),
     getDailyRollup: vi.fn().mockResolvedValue({ days: [], totalSessions: 0 }),
     getProfileOverview: vi.fn().mockResolvedValue(makeProfileOverview()),
     detectInstalledAgents: vi.fn().mockResolvedValue([]),
